more styling, more Interaction components to their own file
