+++ src/components/Notes/Interaction/index.ts
@@ -1,0 +1,1 @@
+export * from "./Interaction";
