+++ src/components/Notes/Notes.css
@@ -2,15 +2,43 @@
   display: flex;
   flex-direction: column;
 
-  .neo-notes__heading {
-    margin-bottom: 1rem;
+  --notes-border-color: var(--neo-color-base-200);
+  @media (prefers-color-scheme: dark) {
+    :root {
+      --notes-border-color: var(--neo-color-base-600);
+    }
   }
 
-  .neo-notes__interaction:not(:last-child) {
-    border-bottom: 1px solid #e0e0e0;
-    padding-bottom: 2rem;
+  .neo-notes__heading {
+    p {
+      font-size: 1.5rem;
+      font-weight: 400;
+      padding-bottom: 1rem;
+    }
+  }
 
-    /* neo-notes__interaction__heading */
+  .neo-notes__interaction {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+    padding-top: 1rem;
+
+    &:not(:last-child) {
+      border-bottom: 1px solid var(--notes-border-color);
+      padding-bottom: 2rem;
+    }
+
+    .neo-notes__interaction__heading {
+      display: flex;
+      align-items: center;
+      font-size: 16px;
+      gap: 0.5rem;
+
+      p {
+        font-weight: 600;
+      }
+    }
+
     /* neo-notes__interaction__message */
     /* neo-notes__interaction__message__title */
     /* neo-notes__interaction__message__content */
