+++ src/components/Notes/Notes.stories.css
@@ -1,0 +1,5 @@
+.notes-container {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+}
